Create my work section html and styling
I added the my work section of the homepage that shows off a set of
projects. Each project has information to describe what it is and links
to go to the github repo for each project. There is just dummy content,
I didn't add any real info. Styling for mobile, tablet, and desktop
implemented.